Split entity operations into modular files
